+++ src/index.ts
@@ -2,7 +2,7 @@ export { buildContractClass, VerifyResult, buildTypeClasses, buildStructsClass, 
 export { compile, StructEntity, getStructDeclaration, getABIDeclaration, ABIEntity, ABIEntityType, ABI, ParamEntity, BuildType, RelatedInformation } from './internal';
 export { Arguments, Argument } from './internal';
 export {
-  bsv, ECIES, signTx, toHex, getPreimage, num2bin, bin2num, bool2Asm, int2Asm, parseLiteral, bytes2Literal, bytesToHexString, getValidatedHexString, literal2ScryptType, literal2Asm,
+  bsv, ECIES, signTx, toHex, getPreimage, bin2num, bool2Asm, int2Asm, parseLiteral, bytes2Literal, bytesToHexString, getValidatedHexString, literal2ScryptType, literal2Asm,
   findStructByType, findStructByName, isStructType, isArrayType, compileContract,
   arrayTypeAndSize, newCall, getStructNameByType, genLaunchConfigFile, subArrayType,
   flattenSha256, isGenericType, parseGenericType, findLibraryByGeneric, toData, findKeyIndex
@@ -12,6 +12,8 @@ export {
   Int, Bool, Bytes, PrivKey, PubKey, Sig, Ripemd160, Sha1, Sha256, SigHashType, SigHashPreimage, OpCodeType, SingletonParamType, SupportedParamType, ScryptType, ValueType,
   Struct, StructObject, VariableType, TypeResolver, serializeSupportedParamType, PrimitiveTypes, RawTypes
 } from './internal';
+
+export { hash160, sha256, hash256, num2bin } from './internal';
 
 export {
   getPlatformScryptc, findCompiler